MVI 구조에 맞게 파일 분리
