+++ Moda/Feature/Chat/List/ChatListIntent.swift
@@ -1,0 +1,14 @@
+//
+//  ChatListIntent.swift
+//  Moda
+//
+//  Created by 금가경 on 11/29/25.
+//
+
+import Foundation
+
+enum ChatListIntent {
+    case chatRoomTapped(ChatRoom)
+    case loadRooms
+    case refresh
+}
