+++ Moda/Feature/Friend/Add/FriendAddView.swift
@@ -8,230 +8,6 @@
 import SwiftUI
 import Observation
 
-private struct FriendSearchItem: Identifiable, Hashable {
-    let id: String          // userId
-    let nickname: String    // nick
-    let profileImageURL: URL?
-}
-
-private struct FriendAddState {
-    var query: String = ""
-    let maxLength: Int = 20
-
-    // 검색 관련 상태
-    var results: [FriendSearchItem] = []
-    var isLoading: Bool = false
-    var errorMessage: String?
-    var hasSearched: Bool = false // 첫 검색 여부
-
-    // 선택된 셀(선택 시 목록 숨기고 카드만 노출)
-    var selectedItem: FriendSearchItem?
-
-    // 선택된 항목의 "친구 여부"
-    // true: 이미 내가 팔로우 중 → 버튼은 "친구 취소"
-    // false 또는 nil: 팔로우 아님(또는 미확정) → 버튼은 "친구 추가"
-    var selectedIsFriend: Bool?
-
-    // 팔로우/언팔로우 진행 상태
-    var isFollowUpdating: Bool = false
-}
-
-
-private enum FriendAddIntent {
-    case onAppear
-    case queryChanged(String)
-    case clearTapped
-    case searchSubmitted
-    case rowTapped(FriendSearchItem)
-
-    // 카드 친구 추가/취소 버튼 탭
-    case friendAddButtonTapped
-
-    // 카드 닫기(X) 버튼 탭
-    case cardCloseTapped
-}
-
-@MainActor
-@Observable
-private final class FriendAddStore {
-    var state = FriendAddState()
-
-    // 의존성
-    private let userAPI = UserAPI.shared
-    private let userProfileAPI: UserProfileAPIProtocol = UserProfileAPI.shared
-    private let followAPI = FollowAPI.shared
-
-    // 동시 검색 취소용
-    private var searchTask: Task<Void, Never>?
-
-    // 내 사용자 ID
-    private var myUserId: String?
-
-    func send(_ intent: FriendAddIntent) {
-        switch intent {
-        case .onAppear:
-            handleOnAppear()
-        case .queryChanged(let text):
-            handleQueryChanged(text)
-        case .clearTapped:
-            state.query = ""
-        case .searchSubmitted:
-            handleSearchSubmitted()
-        case .rowTapped(let item):
-            handleRowTapped(item)
-        case .friendAddButtonTapped:
-            handlefriendAddButtonTapped()
-        case .cardCloseTapped:
-            handleCardCloseTapped()
-        }
-    }
-
-    private func handleOnAppear() {
-        // 최초 1회 내 프로필 로드(내 userId 확보)
-        guard myUserId == nil else { return }
-        Task { [weak self] in
-            guard let self else { return }
-            do {
-                let me = try await userProfileAPI.getMyProfile()
-                self.myUserId = me.userId
-            } catch {
-                // 내 ID 로드 실패는 치명적이지 않으므로 로그만
-            }
-        }
-    }
-
-    private func handleQueryChanged(_ text: String) {
-        // 입력 중 최대길이 도달시 입력 방지
-        state.query = clampToMaxLength(text)
-    }
-
-    private func handleSearchSubmitted() {
-        // 필요 시 공백 제거 + 길이 보정
-        let trimmed = state.query.trimmingCharacters(in: .whitespacesAndNewlines)
-        state.query = clampToMaxLength(trimmed)
-
-        // 빈 입력은 무시
-        guard !state.query.isEmpty else { return }
-
-        // 첫 검색 실행 표시
-        state.hasSearched = true
-
-        // 이전 검색 취소
-        searchTask?.cancel()
-
-        // 선택 상태 초기화(새 검색 시작 시 카드 숨김)
-        state.selectedItem = nil
-        state.selectedIsFriend = nil
-
-        // 로딩 시작
-        state.isLoading = true
-        state.errorMessage = nil
-        state.results = []
-
-        let query = state.query
-
-        // 실제 네트워크 검색
-        searchTask = Task { [weak self] in
-            guard let self else { return }
-            do {
-                let response = try await userAPI.searchUsers(query: query)
-
-                // 취소되었으면 중단
-                if Task.isCancelled { return }
-
-                // FriendListView 방식으로 단순 URL 생성
-                var items: [FriendSearchItem] = response.data.map { dto in
-                    FriendSearchItem(
-                        id: dto.userId,
-                        nickname: dto.nick,
-                        profileImageURL: URL(string: "\(NetworkConfig.baseURL)/v1\(dto.profileImage ?? "")")
-                    )
-                }
-
-                // 내 아이디 제외
-                if let myId = self.myUserId {
-                    items.removeAll { $0.id == myId }
-                }
-
-                state.results = items
-            } catch {
-                if let netErr = error as? NetworkError {
-                    state.errorMessage = netErr.localizedDescription
-                } else {
-                    state.errorMessage = error.localizedDescription
-                }
-            }
-            state.isLoading = false
-        }
-    }
-
-    private func handleRowTapped(_ item: FriendSearchItem) {
-        // 선택된 셀로 카드 표시
-        state.selectedItem = item
-        state.selectedIsFriend = nil // 아직 모르는 상태(nil) → 버튼은 "친구 추가"로 노출
-
-        // 선택한 유저의 상세 프로필 조회하여 followers에 내 ID가 있는지 확인
-        Task { [weak self] in
-            guard let self else { return }
-            do {
-                let profile = try await userProfileAPI.getUserProfile(userId: item.id)
-                if let myId = self.myUserId {
-                    let isFriend = profile.followers.contains { $0.userId == myId }
-                    state.selectedIsFriend = isFriend
-                } else {
-                    // 내 ID를 모르면 판단 불가 → 기본 false
-                    state.selectedIsFriend = false
-                }
-            } catch {
-                if let netErr = error as? NetworkError {
-                    state.errorMessage = netErr.localizedDescription
-                } else {
-                    state.errorMessage = error.localizedDescription
-                }
-                // 판단 실패 시 기본값 유지
-            }
-        }
-    }
-
-    private func handlefriendAddButtonTapped() {
-        guard let selected = state.selectedItem, state.isFollowUpdating == false else { return }
-
-        // 현재 상태 반대로 요청
-        let shouldFollow = !(state.selectedIsFriend ?? false)
-        state.isFollowUpdating = true
-
-        Task { [weak self] in
-            guard let self else { return }
-            do {
-                _ = try await followAPI.follow(userId: selected.id, followStatus: shouldFollow)
-                state.selectedIsFriend = shouldFollow
-            } catch {
-                if let netErr = error as? NetworkError {
-                    state.errorMessage = netErr.localizedDescription
-                } else {
-                    state.errorMessage = error.localizedDescription
-                }
-            }
-            state.isFollowUpdating = false
-        }
-    }
-
-    private func handleCardCloseTapped() {
-        // 카드 닫기: 선택 해제 → 기존 검색 결과 리스트 노출
-        state.selectedItem = nil
-        state.selectedIsFriend = nil
-        state.isFollowUpdating = false
-    }
-
-    private func clampToMaxLength(_ text: String) -> String {
-        if text.count > state.maxLength {
-            return String(text.prefix(state.maxLength))
-        } else {
-            return text
-        }
-    }
-}
-
 struct FriendAddView: View {
     @State private var store = FriendAddStore()
     @FocusState private var isSearching: Bool
