+++ Moda/Feature/Friend/List/FriendListIntent.swift
@@ -1,0 +1,14 @@
+//
+//  FriendListIntent.swift
+//  Moda
+//
+//  Created by 금가경 on 11/29/25.
+//
+
+import Foundation
+
+enum FriendListAction {
+    case onAppear
+    case refresh
+    case dismissError
+}
