+++ Moda/Feature/Friend/List/FriendListView.swift
@@ -8,111 +8,6 @@
 import SwiftUI
 import Observation
 
-// MARK: - Model (View 전용 표시 모델)
-struct People: Identifiable, Hashable {
-    let id: String
-    var name: String
-    var statusMessage: String?
-    var profileImageURL: URL?
-}
-
-// MARK: - State (원본 DTO를 보관)
-struct FriendListState {
-    var myProfile: MyProfileResponse?
-    var friends: [OtherProfileResponse] = []
-    var isLoading: Bool = false
-    var errorMessage: String?
-}
-
-// MARK: - Intent
-enum FriendListAction {
-    case onAppear
-    case refresh
-    case dismissError
-}
-
-// MARK: - Store
-@MainActor
-@Observable
-final class FriendListStore {
-    private(set) var state = FriendListState()
-    private let userProfileAPI: UserProfileAPIProtocol = UserProfileAPI.shared
-
-    // 파생 상태(표시용)
-    var myPeople: People? {
-        guard let my = state.myProfile else { return nil }
-        return People(
-            id: my.userId,
-            name: my.nick,
-            statusMessage: my.info1,
-            profileImageURL: URL(string: "\(NetworkConfig.baseURL)/v1\(my.profileImage ?? "")")
-        )
-    }
-
-    var friendPeople: [People] {
-        state.friends.map { friend in
-            People(
-                id: friend.userId,
-                name: friend.nick,
-                statusMessage: friend.info1,
-                profileImageURL: URL(string: "\(NetworkConfig.baseURL)/v1\(friend.profileImage ?? "")")
-            )
-        }
-    }
-
-    func send(_ action: FriendListAction) {
-        switch action {
-        case .onAppear, .refresh:
-            Task { await load() }
-        case .dismissError:
-            state.errorMessage = nil
-        }
-    }
-
-    // MARK: - Side Effects
-    private func load() async {
-        guard !state.isLoading else { return }
-        state.isLoading = true
-        defer { state.isLoading = false }
-
-        do {
-            // 1) 내 프로필 조회 (DTO 그대로 사용)
-            let dto = try await userProfileAPI.getMyProfile()
-            state.myProfile = dto
-
-            // 2) 맞팔 계산: followers ∩ following (userId 기준)
-            let followerIDs = Set(dto.followers.map { $0.userId })
-            let followingIDs = Set(dto.following.map { $0.userId })
-            let mutualIDs = Array(followerIDs.intersection(followingIDs))
-
-            // 3) 맞팔 사용자 상세 정보 조회(OtherProfileResponse)
-            var details: [OtherProfileResponse] = []
-            for id in mutualIDs {
-                do {
-                    let profile = try await userProfileAPI.getUserProfile(userId: id)
-                    details.append(profile)
-                } catch {
-                    // 개별 친구 상세 조회 실패는 스킵
-                    continue
-                }
-            }
-
-            // 4) 보기 좋게 닉네임 기준 정렬
-            state.friends = details.sorted {
-                $0.nick.localizedCaseInsensitiveCompare($1.nick) == .orderedAscending
-            }
-
-        } catch {
-            if let netErr = error as? NetworkError {
-                state.errorMessage = netErr.localizedDescription
-            } else {
-                state.errorMessage = error.localizedDescription
-            }
-        }
-    }
-}
-
-
 struct FriendListView: View {
     @State private var store = FriendListStore()
     @EnvironmentObject var navigator: AppNavigator
