+++ Moda/Feature/Friend/Search/FriendSearchIntent.swift
@@ -1,0 +1,13 @@
+//
+//  FriendSearchIntent.swift
+//  Moda
+//
+//  Created by 금가경 on 11/29/25.
+//
+
+import Foundation
+
+enum FriendSearchAction {
+    case queryChanged(String)
+    case clearTapped
+}
