+++ Moda/Feature/Friend/Search/FriendSearchState.swift
@@ -1,0 +1,13 @@
+//
+//  FriendSearchState.swift
+//  Moda
+//
+//  Created by 금가경 on 11/29/25.
+//
+
+import Foundation
+
+struct FriendSearchState {
+    var query: String = ""
+    var results: [People] = []
+}
