+++ Moda/Feature/MainTabViewIntent.swift
@@ -1,0 +1,12 @@
+//
+//  MainTabViewIntent.swift
+//  Moda
+//
+//  Created by 금가경 on 11/29/25.
+//
+
+import Foundation
+
+enum MainTabViewIntent {
+    case tabSelected(TabItem)
+}
