+++ Moda/Feature/MainTabViewState.swift
@@ -1,0 +1,12 @@
+//
+//  MainTabViewState.swift
+//  Moda
+//
+//  Created by 금가경 on 11/29/25.
+//
+
+import Foundation
+
+struct MainTabViewState {
+    var selectedTab: TabItem = .home
+}
